Make throttle repositories async, and parallelise lookups.
If your throttle repository is off-machine (as it must be if you want cross-server-farm throttling) the synchronous nature of the repository lookups caused unacceptable performance problems in cases where multiple policies applied, because there would be a series of sequential lookups.

E.g.: query for 1s time window says we're not over quota, so we then query for 1m, and if that says OK, we query for 1h, and if that says ok...and so on. You can end up with lots of queries one after the other. And the problem is that you hit the worst case in the happy path - in order to determine that you're not exceeding any of the quotas, it has to do all of the lookups.

So I've modified it to kick off all the lookups it would need in the happy case simultaneously, and it only starts waiting for results after all the requests are in flight. This way we typically pay the server-to-repository latency cost once, rather than once for each period size.

When testing locally against a cloud-based Redis cache this makes a huge difference - it was taking well over a second to process a request before, but now it takes just a little longer than the time taken for a single request to the cache.