Add support for throttling by user ID.
The code originally just had the idea of a 'client key' which mapped onto a header value, which wasn't a very effective way of distinguishing between users. For one thing, each time the user refreshes their token, the authorization header will change. Also, it provides us no easy way to correlate blocking log reports with particular users.

So I've introduced a distinction between User ID and Client Type Key. A user ID identifies a particular user (and is based on the name ID claim stored in the claims principal; obviously this only works for sites using claims-based auth, but for my present purposes, that's fine). The 'client type key' is really just a renaming of the old approach - a slightly less misleading name. TBH, I'm not sure when the old approach would actually be workable in practice, so at some point we could either remove the old approach, or replace it with some pluggable way where you can supply a strategy for determining what type of client you're dealing with (e.g., guest, admin, free user, paid user etc) and making that the ClientTypeKey.

We also grab the user name (the name the user logs in with, often an email address; the user ID is normally some opaque internal identifier, often a GUID) and store that in log entries, but don't currently use it as a key for throttling. This is simply to make it easier to work out who is running into throttling limits when inspecting logs.