Improve client type handling.
We now have a mechanism for determining client type that doesn't depend on us trusting the client to provide a header that says what kind of client it is. (E.g., "Hey, I'm an admin, honest!".) The app can supply code that decides what the client type key is based either on the claims or the headers.